requirements.txt
fastapi==0.95.2
starlette==0.27.0
uvicorn==0.22.0
sse_starlette==1.1.6
mcp[cli]==1.15.0
flask-socketio==5.3.2
fuzzywuzzy==0.18.0
python-dotenv==1.0.0
httpx>=0.27.1
python-Levenshtein==0.21.0